moved internal methods to package private class
