+++ src/main/java/org/ossgang/commons/mapbackeds/Mapbackeds.java
@@ -4,12 +4,12 @@ import static java.util.Collections.emptyMap;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toMap;
 import static java.util.stream.Collectors.toSet;
+import static org.ossgang.commons.mapbackeds.MapbackedInternals.requireInterface;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -17,8 +17,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 /**
  * This class provides static methods for working with Mapbacked objects. Mapbacked objects are intended to be used e.g.
@@ -245,45 +243,7 @@ public final class Mapbackeds {
     }
 
     public static Set<Method> fieldMethods(Class<?> intfc) {
-        return collectInterfaceMethods(intfc, m -> !m.isDefault());
-    }
-
-    public static Optional<Method> toStringMethod(Class<?> intfc) {
-        Set<Method> toStringMethods = toStringMethods(intfc);
-        if (toStringMethods.size() > 1) {
-            throw new IllegalArgumentException(
-                    "More than one method with a ToString annotation found in the hierarchy!");
-        }
-        return toStringMethods.stream().findFirst();
-    }
-
-    private static Set<Method> toStringMethods(Class<?> intfc) {
-        return collectInterfaceMethods(intfc, m -> m.isAnnotationPresent(ToString.class));
-    }
-
-    private static Set<Method> collectInterfaceMethods(Class<?> intfc, Predicate<Method> methodPredicate) {
-        requireInterface(intfc);
-        Set<Method> fields = findZeroParamMethodsFrom(intfc, methodPredicate);
-        Class<?>[] superinterfaces = intfc.getInterfaces();
-        for (Class<?> supi : superinterfaces) {
-            fields.addAll(collectInterfaceMethods(supi, methodPredicate));
-        }
-        return fields;
-    }
-
-    private static Set<Method> findZeroParamMethodsFrom(Class<?> intfc, Predicate<Method> methodPredicate) {
-        return Arrays.stream(intfc.getDeclaredMethods()) //
-                .filter(m -> m.getParameterCount() == 0) //
-                .filter(m -> methodPredicate.test(m)) //
-                .collect(Collectors.toSet());
-    }
-
-    public static <C> Class<C> requireInterface(Class<C> intfc) {
-        requireNonNull(intfc, "interface must not be null");
-        if (!intfc.isInterface()) {
-            throw new IllegalArgumentException("Given class '" + intfc + "'is not an interface!");
-        }
-        return intfc;
+        return MapbackedInternals.fieldMethods(intfc);
     }
 
     private final static class MethodCapture implements InvocationHandler {
